requirements.txt
snowflake-connector-python==3.7.1
confluent_kafka==2.3.0
python-dotenv==1.0.1